requirements.txt
Django==3.1.7
gunicorn==20.0.4
pyjsgf==1.9.0